EE-16064 Move the ignored postgres JSONB test to its own class
